remove cast

+++ two-bulb-imp/main.cpp
@@ -30,12 +30,12 @@ int main(int argc, const char * argv[]) {
     b_data_t bulb_data;
     bulb_data.mol_fracs_bulb1.x1 = 0.501; // Bulb 1 H2 fraction
     bulb_data.mol_fracs_bulb1.x2 = 0.499; // Bulb 1 N2 fraction
-    bulb_data.mol_fracs_bulb1.x3 = 1 - 0.501 - 0.499; // Bulb 1 CO2 fraction
+    bulb_data.mol_fracs_bulb1.x3 = 1.0 - 0.501 - 0.499; // Bulb 1 CO2 fraction
     
     // Initial composition bulb 2
     bulb_data.mol_fracs_bulb2.x1 = 0.0; // Bulb 2 H2 fraction
     bulb_data.mol_fracs_bulb2.x2 = 0.501; // Bulb 2 N2 fraction
-    bulb_data.mol_fracs_bulb2.x3 = 1 - 0.0 - 0.501; // Bulb 2 CO2 fraction
+    bulb_data.mol_fracs_bulb2.x3 = 1.0 - 0.0 - 0.501; // Bulb 2 CO2 fraction
     
     // Total concentration
     p_params_t p_params;
@@ -46,7 +46,7 @@ int main(int argc, const char * argv[]) {
     t_params.to = 0.0; // Initial time (h)
     t_params.tf = 20.0; // Final time (h)
     t_params.nt = 5; // Number of time steps.
-    t_params.dt = (double) (t_params.tf - t_params.to) / t_params.nt; // Time sampling
+    t_params.dt = (t_params.tf - t_params.to) / t_params.nt; // Time sampling
     
     // Diffusivities
     p_params.D12 = 8.33e-5 * 3600; // units (m2 / h)
